Convert unsigned int to types::global_dof_index in the tests.
git-svn-id: https://svn.dealii.org/branches/branch_bigger_global_dof_indices_4@28961 0785d39b-7218-0410-832d-ea1e28bc413d

+++ tests/bits/fe_tools_common.h
@@ -83,7 +83,8 @@ output_vector (const VECTOR &v)
 
                                    // write out at most 20 equispaced
                                    // elements of the vector
-  for (unsigned int i=0; i<v.size(); i+=std::max(1U,v.size()/20))
+  for (unsigned int i=0; i<v.size(); i+=std::max(static_cast<std::size_t>(1),
+        v.size()/20))
     deallog << v(i) << ' ';
   deallog << std::endl;
 }
